Move favorite content from scenes to components
