+++ src/components/FavoriteDrawer/styles.js
@@ -13,6 +13,19 @@ const styles = (theme) => ({
   drawerUnderNavbar: {
     zIndex: 999,
   },
+  favoriteSidebarPlacholderContainer: {
+    marginTop: theme.padding.space7X.px,
+  },
+  favoriteSidebarPlacholderText: {
+    display: 'grid',
+    marginTop: theme.padding.space3X.px,
+    padding: theme.padding.space5X.px,
+    justifyContent: 'center',
+  },
+  favoriteSidebarPlacholderButton: {
+    display: 'grid',
+    justifyContent: 'center',
+  },
 });
 
 export default styles;
