--- src/scenes/Activities/Favorite/styles.js
@@ -1,17 +1,0 @@
-const styles = (theme) => ({
-  favoriteSidebarPlacholderContainer: {
-    marginTop: theme.padding.space7X.px,
-  },
-  favoriteSidebarPlacholderText: {
-    display: 'grid',
-    marginTop: theme.padding.space3X.px,
-    padding: theme.padding.space5X.px,
-    justifyContent: 'center',
-  },
-  favoriteSidebarPlacholderButton: {
-    display: 'grid',
-    justifyContent: 'center',
-  },
-});
-
-export default styles;
